requirements.txt
python-telegram-bot==21.5
groq==0.9.0
python-dotenv==1.0.1
